fix(posto): confimacao antes de deletar

+++ codigo/public/assets/js/postos.js
@@ -40,8 +40,15 @@ async function init() {
     document.addEventListener("click", (e) => {
     if (e.target.classList.contains("btnDeletar")) {
         const id = e.target.dataset.id;
+        const confirmar = confirm(
+            "Tem certeza que deseja deletar este posto?"
+        );
+        if (!confirmar) {
+            return;
+        }
         console.log("ID do posto deletado:", id);
         deletePosto(id);
+        location.reload();
     }
 });
 }
